requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.1.0
fastjsonschema==2.21.1
fpdf==1.7.2
frozenlist==1.5.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.3
openai==0.27.0
openpyxl==3.1.2
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20200517
pdfplumber==0.5.28
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.21.0
Pygments==2.18.0
PyMuPDF==1.22.5
PyPDF2==3.0.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
setuptools==75.5.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
textdistance==4.6.3
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
tzdata==2024.2
urllib3==2.2.3
Wand==0.6.13
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.45.0
yarg==0.1.9
yarl==1.17.1
